pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "msfabric-devops"
version = "0.1.0"
description = "Python package to interact with Microsoft Fabric items."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Hugo Salaun", email = "hcrsalaun@gmail.com" }
]
dependencies = [
  "azure-identity>=1.17",
  "requests>=2.32"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["msfabric_devops*"]
